+++ stitch/Fragmentation/Spectrum.cs
@@ -16,13 +16,13 @@ namespace Stitch {
         /// <param name="peptides">All peptides to find the spectra for.</param>
         /// <param name="directory">The directory in which to search for the raw data files.</param>
         /// <returns></returns>
-        public static Dictionary<string, List<AnnotatedSpectrumMatch>> GetSpectra(IEnumerable<Read.IRead> peptides, string directory) {
+        public static Dictionary<string, List<AnnotatedSpectrumMatch>> GetSpectra(IEnumerable<Read.IRead> peptides) {
             var fragments = new Dictionary<string, List<AnnotatedSpectrumMatch>>(peptides.Count());
             var scans = peptides.SelectMany(p => p.ScanNumbers.Select(s => (p.EscapedIdentifier, s.RawFile, s.Scan, s.OriginalTag)));
 
             foreach (var group in scans.GroupBy(m => m.RawFile)) {
                 ThermoRawFile raw_file = new ThermoRawFile();
-                var raw_file_path = directory + Path.DirectorySeparatorChar + group.Key;
+                var raw_file_path = group.Key;
                 var correct_path = InputNameSpace.ParseHelper.TestFileExists(raw_file_path);
                 if (!correct_path.IsErr()) {
                     try {